updated navbar, dropdown, catalog, 'finish day' + fixed ps student view
